Rename leaderboard objects module
